app/requirements.txt
google-cloud-pubsub 
google-api-python-client 
PyYAML